+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-24T09:48:33.978Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-24T12:43:22.642Z")
     r"""The date and time the resource was last updated"""