+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8ASXPHZWHNGY43JYCGNWZC3"
+        "01K8B3XS3G9DBP9Y0633Q27Q1N"
     )
     r"""The id of the resource"""