+++ packages/orq-rc/RELEASES.md
@@ -3599,3 +3599,13 @@ Based on:
 - [python v3.14.0-rc.55] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.55] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.55 - packages/orq-rc
+
+## 2025-10-24 12:45:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.639.2 (2.730.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.56] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.56] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.56 - packages/orq-rc